Separation of the files fat_node and nodes_table
